Worksheet text complete, sans images
As title. Text body changed as discussed. Added warning boxes for information which is important (ie students can't complete the worksheet without it). Images in progress. sensing.py and racing.py have variable names changed as discussed to hopefully make it more intuitive / readable for students. Added brief notes in both and referenced them directly in the worksheet to hopefully force students to read them and draw conclusions themselves. Left almost all the jargon in the comments of racing.py as at that point, they should understand the core of how the code "reads" the sensor's values.